fixed to use go1.18
