+++ .github/workflows/test.yml
@@ -1,0 +1,30 @@
+
+name: test
+on:
+  push:
+    branches:
+      - "main"
+    tags:
+      - "v*.*.*"
+  pull_request:
+
+jobs:
+  test:
+    name: Test
+    runs-on: ubuntu-latest
+    steps:
+    - name: Check out code into the Go module directory
+      uses: actions/checkout@v2
+    - uses: actions/setup-go@v2
+      with:
+        go-version: '^1.18'
+    - name: Declare some variables
+      id: vars
+      run: |
+        echo "::set-output name=coverage_txt::${RUNNER_TEMP}/coverage.txt"
+    - name: Test Coverage (pkg)
+      run: go test ./... -race -coverprofile=${{ steps.vars.outputs.coverage_txt }}
+    - name: Upload coverage
+      uses: codecov/codecov-action@v2
+      with:
+        files: ${{ steps.vars.outputs.coverage_txt }}
