+++ go.mod
@@ -1,5 +1,5 @@
 module github.com/Code-Hex/dd
 
-go 1.17
+go 1.18
 
 require github.com/google/go-cmp v0.5.7
